Athena tests refactoring.
