+++ tests/test_athena_csv.py
@@ -1,0 +1,160 @@
+import logging
+
+import pandas as pd
+import pytest
+
+import awswrangler as wr
+
+from ._utils import ensure_data_types_csv, get_df_csv
+
+logging.basicConfig(level=logging.INFO, format="[%(asctime)s][%(levelname)s][%(name)s][%(funcName)s] %(message)s")
+logging.getLogger("awswrangler").setLevel(logging.DEBUG)
+logging.getLogger("botocore.credentials").setLevel(logging.CRITICAL)
+
+
+def test_csv_dataset(path, glue_database):
+    with pytest.raises(wr.exceptions.UndetectedType):
+        wr.s3.to_csv(pd.DataFrame({"A": [None]}), path, dataset=True, database=glue_database, table="test_csv_dataset")
+    df = get_df_csv()
+    with pytest.raises(wr.exceptions.InvalidArgumentCombination):
+        wr.s3.to_csv(df, path, dataset=False, mode="overwrite", database=glue_database, table="test_csv_dataset")
+    with pytest.raises(wr.exceptions.InvalidArgumentCombination):
+        wr.s3.to_csv(df, path, dataset=False, table="test_csv_dataset")
+    with pytest.raises(wr.exceptions.InvalidArgumentCombination):
+        wr.s3.to_csv(df, path, dataset=True, mode="overwrite", database=glue_database)
+    with pytest.raises(wr.exceptions.InvalidArgumentCombination):
+        wr.s3.to_csv(df=df, path=path, mode="append")
+    with pytest.raises(wr.exceptions.InvalidArgumentCombination):
+        wr.s3.to_csv(df=df, path=path, partition_cols=["col2"])
+    with pytest.raises(wr.exceptions.InvalidArgumentCombination):
+        wr.s3.to_csv(df=df, path=path, description="foo")
+    with pytest.raises(wr.exceptions.InvalidArgumentValue):
+        wr.s3.to_csv(df=df, path=path, partition_cols=["col2"], dataset=True, mode="WRONG")
+    paths = wr.s3.to_csv(
+        df=df,
+        path=path,
+        sep="|",
+        index=False,
+        use_threads=True,
+        boto3_session=None,
+        s3_additional_kwargs=None,
+        dataset=True,
+        partition_cols=["par0", "par1"],
+        mode="overwrite",
+    )["paths"]
+    wr.s3.wait_objects_exist(paths=paths)
+    df2 = wr.s3.read_csv(path=paths, sep="|", header=None)
+    assert len(df2.index) == 3
+    assert len(df2.columns) == 8
+    assert df2[0].sum() == 6
+    wr.s3.delete_objects(path=paths)
+
+
+def test_csv_catalog(path, glue_table, glue_database):
+    df = get_df_csv()
+    paths = wr.s3.to_csv(
+        df=df,
+        path=path,
+        sep="\t",
+        index=True,
+        use_threads=True,
+        boto3_session=None,
+        s3_additional_kwargs=None,
+        dataset=True,
+        partition_cols=["par0", "par1"],
+        mode="overwrite",
+        table=glue_table,
+        database=glue_database,
+    )["paths"]
+    wr.s3.wait_objects_exist(paths=paths)
+    df2 = wr.athena.read_sql_table(glue_table, glue_database)
+    assert len(df2.index) == 3
+    assert len(df2.columns) == 11
+    assert df2["id"].sum() == 6
+    ensure_data_types_csv(df2)
+    wr.s3.delete_objects(path=paths)
+    assert wr.catalog.delete_table_if_exists(database=glue_database, table=glue_table) is True
+
+
+def test_csv_catalog_columns(path, glue_database, glue_table):
+    paths = wr.s3.to_csv(
+        df=get_df_csv(),
+        path=path,
+        sep="|",
+        columns=["id", "date", "timestamp", "par0", "par1"],
+        index=False,
+        use_threads=False,
+        boto3_session=None,
+        s3_additional_kwargs=None,
+        dataset=True,
+        partition_cols=["par0", "par1"],
+        mode="overwrite",
+        table=glue_table,
+        database=glue_database,
+    )["paths"]
+    wr.s3.wait_objects_exist(paths=paths)
+    df2 = wr.athena.read_sql_table(glue_table, glue_database)
+    assert len(df2.index) == 3
+    assert len(df2.columns) == 5
+    assert df2["id"].sum() == 6
+    ensure_data_types_csv(df2)
+
+    paths = wr.s3.to_csv(
+        df=pd.DataFrame({"id": [4], "date": [None], "timestamp": [None], "par0": [1], "par1": ["a"]}),
+        path=path,
+        sep="|",
+        index=False,
+        use_threads=False,
+        boto3_session=None,
+        s3_additional_kwargs=None,
+        dataset=True,
+        partition_cols=["par0", "par1"],
+        mode="overwrite_partitions",
+        table=glue_table,
+        database=glue_database,
+    )["paths"]
+    wr.s3.wait_objects_exist(paths=paths)
+    df2 = wr.athena.read_sql_table(glue_table, glue_database)
+    assert len(df2.index) == 3
+    assert len(df2.columns) == 5
+    assert df2["id"].sum() == 9
+    ensure_data_types_csv(df2)
+
+
+def test_athena_csv_types(path, glue_database, glue_table):
+    df = get_df_csv()
+    paths = wr.s3.to_csv(
+        df=df,
+        path=path,
+        sep=",",
+        index=False,
+        use_threads=True,
+        boto3_session=None,
+        s3_additional_kwargs=None,
+        dataset=True,
+        partition_cols=["par0", "par1"],
+        mode="overwrite",
+    )["paths"]
+    wr.s3.wait_objects_exist(paths=paths)
+    columns_types, partitions_types = wr.catalog.extract_athena_types(
+        df=df, index=False, partition_cols=["par0", "par1"], file_format="csv"
+    )
+    wr.catalog.create_csv_table(
+        table=glue_table,
+        database=glue_database,
+        path=path,
+        partitions_types=partitions_types,
+        columns_types=columns_types,
+    )
+    wr.catalog.create_csv_table(
+        database=glue_database, table=glue_table, path=path, columns_types={"col0": "string"}, mode="append"
+    )
+    wr.athena.repair_table(glue_table, glue_database)
+    assert len(wr.catalog.get_csv_partitions(glue_database, glue_table)) == 3
+    df2 = wr.athena.read_sql_table(glue_table, glue_database)
+    assert len(df2.index) == 3
+    assert len(df2.columns) == 10
+    assert df2["id"].sum() == 6
+    ensure_data_types_csv(df2)
+    wr.s3.delete_objects(path=paths)
+    assert wr.catalog.delete_table_if_exists(database=glue_database, table=glue_table) is True
