+++ tests/test_athena_projection.py
@@ -1,0 +1,103 @@
+import logging
+
+import pandas as pd
+
+import awswrangler as wr
+
+from ._utils import dt, ts
+
+logging.basicConfig(level=logging.INFO, format="[%(asctime)s][%(levelname)s][%(name)s][%(funcName)s] %(message)s")
+logging.getLogger("awswrangler").setLevel(logging.DEBUG)
+logging.getLogger("botocore.credentials").setLevel(logging.CRITICAL)
+
+
+def test_to_parquet_projection_integer(glue_database, glue_table, path):
+    df = pd.DataFrame({"c0": [0, 1, 2], "c1": [0, 1, 2], "c2": [0, 100, 200], "c3": [0, 1, 2]})
+    paths = wr.s3.to_parquet(
+        df=df,
+        path=path,
+        dataset=True,
+        database=glue_database,
+        table=glue_table,
+        partition_cols=["c1", "c2", "c3"],
+        regular_partitions=False,
+        projection_enabled=True,
+        projection_types={"c1": "integer", "c2": "integer", "c3": "integer"},
+        projection_ranges={"c1": "0,2", "c2": "0,200", "c3": "0,2"},
+        projection_intervals={"c2": "100"},
+        projection_digits={"c3": "1"},
+    )["paths"]
+    wr.s3.wait_objects_exist(paths=paths, use_threads=False)
+    df2 = wr.athena.read_sql_table(glue_table, glue_database)
+    assert df.shape == df2.shape
+    assert df.c0.sum() == df2.c0.sum()
+    assert df.c1.sum() == df2.c1.sum()
+    assert df.c2.sum() == df2.c2.sum()
+    assert df.c3.sum() == df2.c3.sum()
+
+
+def test_to_parquet_projection_enum(glue_database, glue_table, path):
+    df = pd.DataFrame({"c0": [0, 1, 2], "c1": [1, 2, 3], "c2": ["foo", "boo", "bar"]})
+    paths = wr.s3.to_parquet(
+        df=df,
+        path=path,
+        dataset=True,
+        database=glue_database,
+        table=glue_table,
+        partition_cols=["c1", "c2"],
+        regular_partitions=False,
+        projection_enabled=True,
+        projection_types={"c1": "enum", "c2": "enum"},
+        projection_values={"c1": "1,2,3", "c2": "foo,boo,bar"},
+    )["paths"]
+    wr.s3.wait_objects_exist(paths=paths, use_threads=False)
+    df2 = wr.athena.read_sql_table(glue_table, glue_database)
+    assert df.shape == df2.shape
+    assert df.c0.sum() == df2.c0.sum()
+    assert df.c1.sum() == df2.c1.sum()
+
+
+def test_to_parquet_projection_date(glue_database, glue_table, path):
+    df = pd.DataFrame(
+        {
+            "c0": [0, 1, 2],
+            "c1": [dt("2020-01-01"), dt("2020-01-02"), dt("2020-01-03")],
+            "c2": [ts("2020-01-01 01:01:01.0"), ts("2020-01-01 01:01:02.0"), ts("2020-01-01 01:01:03.0")],
+        }
+    )
+    paths = wr.s3.to_parquet(
+        df=df,
+        path=path,
+        dataset=True,
+        database=glue_database,
+        table=glue_table,
+        partition_cols=["c1", "c2"],
+        regular_partitions=False,
+        projection_enabled=True,
+        projection_types={"c1": "date", "c2": "date"},
+        projection_ranges={"c1": "2020-01-01,2020-01-03", "c2": "2020-01-01 01:01:00,2020-01-01 01:01:03"},
+    )["paths"]
+    wr.s3.wait_objects_exist(paths=paths, use_threads=False)
+    df2 = wr.athena.read_sql_table(glue_table, glue_database)
+    print(df2)
+    assert df.shape == df2.shape
+    assert df.c0.sum() == df2.c0.sum()
+
+
+def test_to_parquet_projection_injected(glue_database, glue_table, path):
+    df = pd.DataFrame({"c0": [0, 1, 2], "c1": ["foo", "boo", "bar"], "c2": ["0", "1", "2"]})
+    paths = wr.s3.to_parquet(
+        df=df,
+        path=path,
+        dataset=True,
+        database=glue_database,
+        table=glue_table,
+        partition_cols=["c1", "c2"],
+        regular_partitions=False,
+        projection_enabled=True,
+        projection_types={"c1": "injected", "c2": "injected"},
+    )["paths"]
+    wr.s3.wait_objects_exist(paths=paths, use_threads=False)
+    df2 = wr.athena.read_sql_query(f"SELECT * FROM {glue_table} WHERE c1='foo' AND c2='0'", glue_database)
+    assert df2.shape == (1, 3)
+    assert df2.c0.iloc[0] == 0
